Added drilldown sensor selection list
Added a reactive drilldown list for the sensor selection that updates in the UI when a catchment is selected.

TODO: Need to create a static dataframe of all catchment - sensor combinations as this currently will only show the sensors present in the data (i.e. selected date range); we want all sensors to be show during a selected time period/catchment when defaulting to "All" sensors regardless if there is event data or not.