Update app.R
Improved functionality of the reset button

TODO:  I think I will adjust the dashoboard to be less reactive and more event driven in the sense of having a "report" button that will rely more on bespoke API queries based upon the inputs and less server side data filtering. This should also improves latency times for all but the largest of requests.